Delete awkward tests that don't improve coverage.
The flakiness of these tests was resolved by the preceding commit, but
in a way that was awkward because it relied on guards checking the
thread they were executing in. This worked reliably for now, but
violates the contract of GuardedExecutor guards, which should only
depend on state that is guarded by the executor.

It turns out that not a single line or branch lost its coverage without
these tests, so it seems best to simply delete them.

[#3]